requirements.txt
Django==3.0.7
Pillow==8.0.1
django-crispy-forms==1.9.2
django-simple-captcha==0.5.12